+++ twilio/rest/conversations/v1/conversation/__init__.py
@@ -670,6 +670,9 @@ class ConversationList(ListResource):
 
     def stream(
         self,
+        start_date: Union[str, object] = values.unset,
+        end_date: Union[str, object] = values.unset,
+        state: Union["ConversationInstance.State", object] = values.unset,
         limit: Optional[int] = None,
         page_size: Optional[int] = None,
     ) -> Iterator[ConversationInstance]:
@@ -679,6 +682,9 @@ class ConversationList(ListResource):
         is reached.
         The results are returned as a generator, so this operation is memory efficient.
 
+        :param str start_date: Start date in ISO8601 format for sorting and filtering list of Conversations.
+        :param str end_date: End date in ISO8601 format for sorting and filtering list of Conversations.
+        :param &quot;ConversationInstance.State&quot; state: State for sorting and filtering list of Conversations. Can be `active`, `inactive` or `closed`
         :param limit: Upper limit for the number of records to return. stream()
                       guarantees to never return more than limit.  Default is no limit
         :param page_size: Number of records to fetch per request, when not set will use
@@ -689,12 +695,20 @@ class ConversationList(ListResource):
         :returns: Generator that will yield up to limit results
         """
         limits = self._version.read_limits(limit, page_size)
-        page = self.page(page_size=limits["page_size"])
+        page = self.page(
+            start_date=start_date,
+            end_date=end_date,
+            state=state,
+            page_size=limits["page_size"],
+        )
 
         return self._version.stream(page, limits["limit"])
 
     async def stream_async(
         self,
+        start_date: Union[str, object] = values.unset,
+        end_date: Union[str, object] = values.unset,
+        state: Union["ConversationInstance.State", object] = values.unset,
         limit: Optional[int] = None,
         page_size: Optional[int] = None,
     ) -> AsyncIterator[ConversationInstance]:
@@ -704,6 +718,9 @@ class ConversationList(ListResource):
         is reached.
         The results are returned as a generator, so this operation is memory efficient.
 
+        :param str start_date: Start date in ISO8601 format for sorting and filtering list of Conversations.
+        :param str end_date: End date in ISO8601 format for sorting and filtering list of Conversations.
+        :param &quot;ConversationInstance.State&quot; state: State for sorting and filtering list of Conversations. Can be `active`, `inactive` or `closed`
         :param limit: Upper limit for the number of records to return. stream()
                       guarantees to never return more than limit.  Default is no limit
         :param page_size: Number of records to fetch per request, when not set will use
@@ -714,12 +731,20 @@ class ConversationList(ListResource):
         :returns: Generator that will yield up to limit results
         """
         limits = self._version.read_limits(limit, page_size)
-        page = await self.page_async(page_size=limits["page_size"])
+        page = await self.page_async(
+            start_date=start_date,
+            end_date=end_date,
+            state=state,
+            page_size=limits["page_size"],
+        )
 
         return self._version.stream_async(page, limits["limit"])
 
     def list(
         self,
+        start_date: Union[str, object] = values.unset,
+        end_date: Union[str, object] = values.unset,
+        state: Union["ConversationInstance.State", object] = values.unset,
         limit: Optional[int] = None,
         page_size: Optional[int] = None,
     ) -> List[ConversationInstance]:
@@ -728,6 +753,9 @@ class ConversationList(ListResource):
         Unlike stream(), this operation is eager and will load `limit` records into
         memory before returning.
 
+        :param str start_date: Start date in ISO8601 format for sorting and filtering list of Conversations.
+        :param str end_date: End date in ISO8601 format for sorting and filtering list of Conversations.
+        :param &quot;ConversationInstance.State&quot; state: State for sorting and filtering list of Conversations. Can be `active`, `inactive` or `closed`
         :param limit: Upper limit for the number of records to return. list() guarantees
                       never to return more than limit.  Default is no limit
         :param page_size: Number of records to fetch per request, when not set will use
@@ -739,6 +767,9 @@ class ConversationList(ListResource):
         """
         return list(
             self.stream(
+                start_date=start_date,
+                end_date=end_date,
+                state=state,
                 limit=limit,
                 page_size=page_size,
             )
@@ -746,6 +777,9 @@ class ConversationList(ListResource):
 
     async def list_async(
         self,
+        start_date: Union[str, object] = values.unset,
+        end_date: Union[str, object] = values.unset,
+        state: Union["ConversationInstance.State", object] = values.unset,
         limit: Optional[int] = None,
         page_size: Optional[int] = None,
     ) -> List[ConversationInstance]:
@@ -754,6 +788,9 @@ class ConversationList(ListResource):
         Unlike stream(), this operation is eager and will load `limit` records into
         memory before returning.
 
+        :param str start_date: Start date in ISO8601 format for sorting and filtering list of Conversations.
+        :param str end_date: End date in ISO8601 format for sorting and filtering list of Conversations.
+        :param &quot;ConversationInstance.State&quot; state: State for sorting and filtering list of Conversations. Can be `active`, `inactive` or `closed`
         :param limit: Upper limit for the number of records to return. list() guarantees
                       never to return more than limit.  Default is no limit
         :param page_size: Number of records to fetch per request, when not set will use
@@ -766,6 +803,9 @@ class ConversationList(ListResource):
         return [
             record
             async for record in await self.stream_async(
+                start_date=start_date,
+                end_date=end_date,
+                state=state,
                 limit=limit,
                 page_size=page_size,
             )
@@ -773,6 +813,9 @@ class ConversationList(ListResource):
 
     def page(
         self,
+        start_date: Union[str, object] = values.unset,
+        end_date: Union[str, object] = values.unset,
+        state: Union["ConversationInstance.State", object] = values.unset,
         page_token: Union[str, object] = values.unset,
         page_number: Union[int, object] = values.unset,
         page_size: Union[int, object] = values.unset,
@@ -781,6 +824,9 @@ class ConversationList(ListResource):
         Retrieve a single page of ConversationInstance records from the API.
         Request is executed immediately
 
+        :param start_date: Start date in ISO8601 format for sorting and filtering list of Conversations.
+        :param end_date: End date in ISO8601 format for sorting and filtering list of Conversations.
+        :param state: State for sorting and filtering list of Conversations. Can be `active`, `inactive` or `closed`
         :param page_token: PageToken provided by the API
         :param page_number: Page Number, this value is simply for client state
         :param page_size: Number of records to return, defaults to 50
@@ -789,6 +835,9 @@ class ConversationList(ListResource):
         """
         data = values.of(
             {
+                "StartDate": start_date,
+                "EndDate": end_date,
+                "State": state,
                 "PageToken": page_token,
                 "Page": page_number,
                 "PageSize": page_size,
@@ -800,6 +849,9 @@ class ConversationList(ListResource):
 
     async def page_async(
         self,
+        start_date: Union[str, object] = values.unset,
+        end_date: Union[str, object] = values.unset,
+        state: Union["ConversationInstance.State", object] = values.unset,
         page_token: Union[str, object] = values.unset,
         page_number: Union[int, object] = values.unset,
         page_size: Union[int, object] = values.unset,
@@ -808,6 +860,9 @@ class ConversationList(ListResource):
         Asynchronously retrieve a single page of ConversationInstance records from the API.
         Request is executed immediately
 
+        :param start_date: Start date in ISO8601 format for sorting and filtering list of Conversations.
+        :param end_date: End date in ISO8601 format for sorting and filtering list of Conversations.
+        :param state: State for sorting and filtering list of Conversations. Can be `active`, `inactive` or `closed`
         :param page_token: PageToken provided by the API
         :param page_number: Page Number, this value is simply for client state
         :param page_size: Number of records to return, defaults to 50
@@ -816,6 +871,9 @@ class ConversationList(ListResource):
         """
         data = values.of(
             {
+                "StartDate": start_date,
+                "EndDate": end_date,
+                "State": state,
                 "PageToken": page_token,
                 "Page": page_number,
                 "PageSize": page_size,