+++ twilio/rest/messaging/v1/service/us_app_to_person.py
@@ -32,7 +32,7 @@ class UsAppToPersonInstance(InstanceResource):
     :ivar messaging_service_sid: The SID of the [Messaging Service](https://www.twilio.com/docs/messaging/services/api) that the resource is associated with.
     :ivar description: A short description of what this SMS campaign does. Min length: 40 characters. Max length: 4096 characters.
     :ivar message_samples: Message samples, at least 1 and up to 5 sample messages (at least 2 for starter/sole proprietor), >=20 chars, <=1024 chars each.
-    :ivar us_app_to_person_usecase: A2P Campaign Use Case. Examples: [ 2FA, EMERGENCY, MARKETING, STARTER...]. STARTER campaign use cases can only be created by STARTER Brands, and there can only be one STARTER campaign created per STARTER Brand.
+    :ivar us_app_to_person_usecase: A2P Campaign Use Case. Examples: [ 2FA, EMERGENCY, MARKETING, SOLE_PROPRIETOR...]. SOLE_PROPRIETOR campaign use cases can only be created by SOLE_PROPRIETOR Brands, and there can only be one SOLE_PROPRIETOR campaign created per SOLE_PROPRIETOR Brand.
     :ivar has_embedded_links: Indicate that this SMS campaign will send messages that contain links.
     :ivar has_embedded_phone: Indicates that this SMS campaign will send messages that contain phone numbers.
     :ivar campaign_status: Campaign status. Examples: IN_PROGRESS, VERIFIED, FAILED.